requirements.txt
boto==2.6.0
fabric==1.5.1
PyYAML==3.10
